Moved metrics.py outside and removed keras metrics
